Allow pandas>2 dependency

+++ causallib/simulation/CausalSimulator3.py
@@ -441,7 +441,7 @@ class CausalSimulator3(object):
 
         # generate latent continuous covariates - every variable is guaranteed to have a population variance of 1.0
         # X_latent = pd.DataFrame(index=patients_index, columns=self.var_types.index)
-        X = pd.DataFrame(index=patients_index, columns=self.var_types.index)
+        X = pd.DataFrame(index=patients_index, columns=self.var_types.index, dtype=float)
         if X_given is not None:  # if a dataset is given, integrate it to the current dataset being build.
             X.loc[:, X_given.columns] = X_given
             for col in X_given.columns:
@@ -1342,7 +1342,7 @@ class CausalSimulator3(object):
             beta = pd.DataFrame(data=np.random.normal(loc=0.0, scale=4.0, size=(degree, X_parents.columns.size)),
                                 columns=X_parents.columns, index=np.arange(degree))
 
-        result_polynomial = pd.DataFrame(data=None, index=X_parents.index, columns=X_parents.columns)
+        result_polynomial = pd.DataFrame(data=None, index=X_parents.index, columns=X_parents.columns, dtype=float)
         degrees = beta.index.to_series()
         # Apply a polynomial to every parent variable
         for var_name, col in X_parents.items():
